+++ cloudflare/kv.go
@@ -6,6 +6,8 @@ import (
 	"io"
 	"syscall/js"
 
+	"github.com/syumai/workers/cloudflare/internal/cfruntimecontext"
+
 	"github.com/syumai/workers/internal/jsutil"
 )
 
@@ -21,7 +23,7 @@ type KVNamespace struct {
 //   - if the given variable name doesn't exist on runtime context, returns error.
 //   - This function panics when a runtime context is not found.
 func NewKVNamespace(ctx context.Context, varName string) (*KVNamespace, error) {
-	inst := getRuntimeContextEnv(ctx).Get(varName)
+	inst := cfruntimecontext.GetRuntimeContextEnv(ctx).Get(varName)
 	if inst.IsUndefined() {
 		return nil, fmt.Errorf("%s is undefined", varName)
 	}
